Order.Services/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.5.3
python-multipart==0.0.6
email-validator==2.1.0
httpx==0.26.0
alembic==1.13.1
